[change] Build cancellation | #BAZEL-309 Done Make sure all subprocesses spawned by bazel-bsp are killed on exit
Build cancellation

Make sure that the Bazel Build execution is stopped once
a json-rpc "cancelRequest" message is issued. LSP4J
handles it in that way, it just cancels
CompletableFutures that it gets from API implementations
(for example BspServerApi). It is our responsibility then
to make sure that once this future is canceled, all the
external processes we spawned are killed as well.

Merge-request: BAZEL-MR-141
Merged-by: Tomasz Pasternak <Tomasz.Pasternak@jetbrains.com>