[feature] Resolve Scala SDK independently for each target (#552)
* Pre-change refactoring

* Resolve Scala SDK independently for each target

This prepares us for changes in `rules_scala` that will allow customizing the Scala version for each target.
In order to achieve that, we can no longer resolve the Scala SDK globally as the maximal version used. We need to use per-target info.
Scala SDK will be still discovered based on the compiler class path.
Now though we will look into an implicit dependency of each target, the `_scalac`.

This change is backward-compatible, as the mentioned data is already available.
It is also forward-compatible with the anticipated cross-build feature of `rules_scala`.
(see: https://github.com/bazelbuild/rules_scala/issues/1290)

The aspect will produce additional data – namely few compiler classpath jars per Scala target.
Perhaps we could review this change in the future to normalize the data back. Now it's not possible, as the data about alternative configurations (here: of scalac) is discarded in the server.

* Cleanup after

* Output "external-deps-resolve" again

* Code style fix

* Remove `_scalac` again from compile deps