improvement: Sort deterministcly if more configurations detected
Previously, we would choose just use first target info, which could be the default config or not. And if the custom config was not being built by default, we would never get artifacts for that produced.

Now, we always have the same stable order, which as far as I've been testing is starting with the default as if we have multiple JVM targets one is usually a copy of another with longer paths. Though there is no way for sure to know which one is the default.

Alternative I was thinking about was trying to get the shortest output path.

We avoid needed to add support for configurations at this point, which is a much bigger task and would require scoping with the bigger bazel team.